requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.6
Flask==3.1.3
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
requests==2.33.1
requests-oauthlib==2.0.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.8
